I'm low and earnestly upset but, I'm a freakin' professional and this is getting done.
it's already almost there.

we're gonna get the retrofit response, we're gonna make a local temp mutablelist, from there you can add to favorites which will be stored in the db -- so isFavorite is kind of pointless now, but whatever.
